ofh: bake fs-offset=16 compensation into BFP decompression exponent
Apply a +8 exponent shift (x256 amplitude, +48.16 dB) to the per-PRB
scaler in all BFP decompression paths (generic, AVX2, AVX512, NEON):

    1 << exponent  ->  1 << (exponent + 8)

Widen the generic scaler from int16_t to int32_t and the
quantizer::to_brain_float in_scale parameter (decompress-only users)
from int16_t to int32_t, since exponent+8 can reach 15 and 1 << 15
does not fit in int16_t. Compression paths are unchanged.

+++ lib/ofh/compression/quantizer.h
@@ -97,7 +97,7 @@ public:
   ///
   /// \remark The size of \c x must be twice the size of \c z as \x is comprised by the quantized pairs of real and
   /// imaginary parts of complex values.
-  void to_brain_float(span<cbf16_t> z, span<const int16_t> x, int16_t in_scale) const
+  void to_brain_float(span<cbf16_t> z, span<const int16_t> x, int32_t in_scale) const
   {
     float scale = gain / in_scale;
     ocuduvec::convert(z, x, scale);
